Compute better popularity values
Previously the popularity value of a tweet was always the sum
of retweets and favorites. Now, we consider how much time has
passed since the tweet was first created. Starting popularity
is always 0. Changes popularity value from an int to a float

+++ roosevelt/web/models/tweet_list.ex
@@ -38,8 +38,8 @@ defmodule Roosevelt.TweetList do
   defp parse_json(json), do: Poison.Parser.parse! json
 
   defp less_popular?(tweet_one, tweet_two) do
-    {popularity_one, _} = Integer.parse tweet_one["popularity"]
-    {popularity_two, _} = Integer.parse tweet_two["popularity"]
+    {popularity_one, _} = Float.parse tweet_one["popularity"]
+    {popularity_two, _} = Float.parse tweet_two["popularity"]
     popularity_one < popularity_two
   end
 end
